requirements.txt
PyYAML==6.0
peewee==3.14.9
psycopg2-binary==2.9.3
PyMySQL==1.0.2